function-fastapi/requirements.txt
azure-functions~=1.12.0
fastapi~=0.88.0
